Add files via upload
The Portfolio Project landing page is a webpage that presents information about the project without requiring the visitor to signup, login, or even use the application. It is possible that the Landing Page is also the main webpage for the project.

Let’s take a closer look at what makes a great landing page.